requirements.txt
findspark==2.0.1
delta-spark==2.4.0
sqlparse==0.4.4; python_version >= '3.5'
python-dotenv==1.0.0; python_version >= '3.8'
requests==2.31.0
pytz==2023.3.post1
deltalake==0.15.1
pandas==2.2.0
duckdb==0.9.2
tabulate==0.9.0